# docs/openapi.yaml
openapi: 3.1.0
info:
  title: gigastt REST API
  version: "1.0"
  description: |
    Offline and streaming speech-to-text REST API powered by GigaAM v3 (rnnt head by default).
    On-device Russian speech recognition via ONNX Runtime.

    ## Endpoints

    - **Health & discovery**: `GET /health`, `GET /ready`, `GET /v1/models`
    - **File transcription**: `POST /v1/transcribe` — upload audio, receive full transcript
    - **Streaming transcription**: `POST /v1/transcribe/stream` — upload audio, receive SSE stream of partial/final segments
    - **OpenAI-compatible**: `POST /v1/audio/transcriptions` — multipart `file` + `model` → `{"text":"..."}`
    - **Async jobs** (opt-in `--enable-jobs`): `POST /v1/jobs`, `GET/DELETE /v1/jobs/{id}`, result + SSE events
    - **Admin**: `POST /v1/admin/reload` — loopback-only model hot-reload
    - **WebSocket** (see `docs/asyncapi.yaml`): `GET /v1/ws`

    Prometheus metrics are optional (behind `--metrics`) and exposed on a
    separate listener (default `127.0.0.1:9090`, override via `--metrics-listen`),
    not on this API surface.

    ## Audio Formats

    Supported upload formats: WAV (PCM, IEEE float, G.711, G.722, GSM 06.10,
    MS/IMA ADPCM, RF64/RIFX/BW64/Wave64), MP3, M4A/AAC, OGG/Vorbis, OGG/Opus (`.opus`),
    WebM/Opus (what a browser's `MediaRecorder` produces), FLAC. Raw headerless
    telephony streams are accepted with
    `?codec=pcmu|pcma|g722&sample_rate=N` (see the transcribe parameters).
...

    ## Limits

...

...

...

...

...
    {
...

...
servers:
...
paths:
...
  /v1/jobs/{id}:
    get:
      summary: Poll job status and progress
      description: |
        Returns the current status of an asynchronous transcription job. When the
        job API is disabled, this path returns `404`.
      operationId: getJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      summary: Cancel a queued or processing job
      description: |
        Cancels a job that is still `queued` or `processing`. Cancelling a queued
        job immediately frees its slot; cancelling a processing job stops result
        propagation but the in-flight triplet is released only when the worker
        finishes. When the job API is disabled, this path returns `404`.
      operationId: cancelJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Job cancelled
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Job cannot be cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Job cannot be cancelled"
                code: job_not_cancellable
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
...
  schemas:
...
    WordInfo:
      type: object
      required:
        - word
        - start
        - end
        - confidence
      properties:
        word:
          type: string
          example: "привет"
          description: Recognized word
        start:
          type: number
          format: double
          example: 0.5
          description: Word start time in seconds
        end:
          type: number
          format: double
          example: 0.8
          description: Word end time in seconds
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          example: 0.97
          description: Softmax confidence score (0.0–1.0), averaged over the word's constituent tokens
        speaker:
          type: integer
          description: Zero-based speaker index (only present when channel split or diarization produced speaker labels)
...
    JobStatusResponse:
      type: object
      required:
        - job_id
        - status
        - processed_seconds
        - percent
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [queued, processing, done, failed, cancelled]
        processed_seconds:
          type: number
          example: 12.5
          description: Seconds of audio considered processed so far
        percent:
          type: integer
          example: 42
          description: Approximate completion percentage (0-100)
        error:
          type: string
          example: "Transcription failed."
          description: Sanitized error message for failed jobs
        partial:
          $ref: "#/components/schemas/PartialTranscript"
    PartialTranscript:
      type: object
      description: |
        Last provisional transcript, retained after cancellation or failure.
        May include an incomplete last word and lacks final post-processing.
      required: [text, words, is_final, timestamp]
      properties:
        text:
          type: string
        words:
          type: array
          items:
            $ref: "#/components/schemas/WordInfo"
        is_final:
          type: boolean
          enum: [false]
        timestamp:
          type: number
        confidence:
          type: number
...
    ErrorResponse:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          example: "Request body exceeds the configured size limit"
          description: Human-readable error description
        code:
          type: string
          example: payload_too_large
          description: Machine-readable error code
        partial:
          $ref: "#/components/schemas/PartialTranscript"
        retry_after_ms:
          type: integer
          example: 30000
          description: |
            Suggested retry delay in milliseconds. Present only for transient
            backpressure errors (pool saturation).
  responses:
...
    ServiceUnavailable:
      description: Server busy or shutting down
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retry
          example: 30
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            timeout:
              summary: Pool saturated
              value:
                error: "Server busy, try again later"
                code: timeout
                retry_after_ms: 30000
            pool_closed:
              summary: Graceful shutdown in progress
              value:
                error: "Server is shutting down"
                code: pool_closed
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            job_not_found:
              summary: Unknown job id
              value:
                error: "Job not found"
                code: job_not_found
            # Note: when the server runs without `--enable-jobs` the /v1/jobs
            # routes are not registered at all, so the response is a bare 404
            # with no JSON body — not this payload.